requirements.txt
openai
pydub
python-dotenv
httpx 